requirements.txt
psycopg2-binary==2.9.7
Flask==2.2.5
gunicorn==23.0.0